# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinanceFlow API",
        "description": "API for managing invoices, customers, and user settings for the FinanceFlow application.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get all customers",
                "operationId": "f4bd4899e4f693a3824fdfdb180a1483",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Fuzzy search term for customers.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startDate",
                        "in": "query",
                        "description": "Filter customers created from this date.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "description": "Filter customers created up to this date.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of customers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "operationId": "6aaaa0c0184102ad9017bed9db7c19ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Get all invoices",
                "operationId": "f6bfbb2f42239b1a42dd1bf248dbfe38",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Fuzzy search term for invoices.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startDate",
                        "in": "query",
                        "description": "Filter invoices with due date from this date.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "description": "Filter invoices with due date up to this date.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of invoices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Invoice"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Create a new invoice",
                "operationId": "af4a99f100c05a7b2326ffe4a60b202c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoiceInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invoice created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoices/{invoiceNo}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Get an invoice by number",
                "operationId": "da95774d6857a1170c33a08a29f2f91c",
                "parameters": [
                    {
                        "name": "invoiceNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invoice not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Update an invoice",
                "operationId": "1181852bc1d7a108874fddd5ed91e6a4",
                "parameters": [
                    {
                        "name": "invoiceNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoiceUpdateInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invoice not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Delete an invoice",
                "operationId": "ae9f5d0e68418d8a9d4f00ef28eaa76f",
                "parameters": [
                    {
                        "name": "invoiceNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Invoice deleted"
                    },
                    "404": {
                        "description": "Invoice not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Customer": {
                "properties": {
                    "customerType": {
                        "type": "string",
                        "enum": [
                            "Bireysel",
                            "Kurumsal"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date"
                    },
                    "tckn": {
                        "type": "string"
                    },
                    "vkn": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CustomerInput": {
                "required": [
                    "customerType",
                    "name",
                    "email",
                    "phone",
                    "address"
                ],
                "properties": {
                    "customerType": {
                        "type": "string",
                        "enum": [
                            "Bireysel",
                            "Kurumsal"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "tckn": {
                        "type": "string"
                    },
                    "vkn": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Invoice": {
                "properties": {
                    "invoiceNo": {
                        "type": "string"
                    },
                    "customerKey": {
                        "type": "string",
                        "format": "email"
                    },
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "dueDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "paymentType": {
                        "type": "string",
                        "enum": [
                            "Credit Card",
                            "PayPal",
                            "Bank Transfer"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Pending",
                            "Paid",
                            "Overdue"
                        ]
                    }
                },
                "type": "object"
            },
            "InvoiceInput": {
                "required": [
                    "customerKey",
                    "invoiceNo",
                    "amount",
                    "dueDate",
                    "paymentType"
                ],
                "properties": {
                    "customerKey": {
                        "type": "string",
                        "format": "email"
                    },
                    "invoiceNo": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "dueDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "paymentType": {
                        "type": "string",
                        "enum": [
                            "Credit Card",
                            "PayPal",
                            "Bank Transfer"
                        ]
                    }
                },
                "type": "object"
            },
            "InvoiceUpdateInput": {
                "required": [
                    "customerKey",
                    "amount",
                    "dueDate",
                    "paymentType"
                ],
                "properties": {
                    "customerKey": {
                        "type": "string",
                        "format": "email"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "dueDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "paymentType": {
                        "type": "string",
                        "enum": [
                            "Credit Card",
                            "PayPal",
                            "Bank Transfer"
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}